+++ src/cpu/mod.rs
@@ -1,2 +1,8 @@
-pub mod instructions;
-pub use instructions::{Bytecode, Immediate, Instruction, Opcode};
+pub mod instruction;
+pub use instruction::Instruction;
+pub mod opcode;
+pub use opcode::Opcode;
+pub mod processor;
+pub use processor::Processor;
+pub mod immediate;
+pub use immediate::{Decoding, Encoding, Immediate};
